Test mesh parallelization

+++ smart/mesh_tools.py
@@ -107,16 +107,25 @@ def facet_topology(f: d.Facet, mf3: d.MeshFunction):
     'internal', or 'interface' (boundary of inner cube)
     """
     # cells adjacent face
-    localCells = [mf3.array()[c.index()] for c in d.cells(f)]
-    if len(localCells) == 1:
-        topology = "boundary"  # boundary facet
-    elif len(localCells) == 2 and localCells[0] == localCells[1]:
-        topology = "internal"  # internal facet
-    elif len(localCells) == 2:
-        topology = "interface"  # interface facet
+
+    # Initialize facet to cell connectivity
+    mf3.mesh().init(2, 3)
+    cell_values = mf3.array()
+    connected_cells = f.entities(f.mesh().topology().dim())
+    if f.exterior():
+        return "boundary", cell_values[connected_cells]
     else:
-        raise Exception("Facet has more than two cells")
-    return (topology, localCells)
+        if len(connected_cells) == 2:
+            if cell_values[connected_cells[0]] == cell_values[connected_cells[1]]:
+                return "internal", cell_values[connected_cells]
+            else:
+                return "interface", cell_values[connected_cells]
+        else:
+            assert len(connected_cells) == 1
+            raise Exception(
+                "Missing information for determining if facet is internal or interface",
+                ", have you turned ghost mode to 'shared_facet'?",
+            )
 
 
 def cube_condition(cell, xmin=0.3, xmax=0.7):
@@ -144,18 +153,19 @@ def create_cubes(N=16, condition=cube_condition):
     10 - Boundary of subvolume 1
     0  - Interior facets
     """
-    # Create a mesh
+    # Create a mesh (use ghost mode to fix interfaces correctly)
+    d.parameters["ghost_mode"] = "shared_facet"
     mesh = d.UnitCubeMesh(N, N, N)
     # Initialize mesh functions
     mf3 = d.MeshFunction("size_t", mesh, 3, 0)
     mf2 = d.MeshFunction("size_t", mesh, 2, 0)
 
     # Mark all cells that satisfy condition as 3, else 1
-    for c in d.cells(mesh):
+    for c in d.cells(mesh, "all"):
         mf3[c] = 2 if condition(c) else 1
 
     # Mark facets
-    for f in d.faces(mesh):
+    for f in d.facets(mesh):
         topology, cellIndices = facet_topology(f, mf3)
         if topology == "interface":
             mf2[f] = 12
@@ -233,62 +243,6 @@ def create_ellipsoids(
     Returns:
         Tuple (mesh, facet_marker, cell_marker)
     """
-    import gmsh
-
-    if np.any(np.isclose(outerRad, 0)):
-        ValueError("One of the outer radii is equal to zero")
-    if np.isclose(hEdge, 0):
-        hEdge = 0.1 * max(outerRad)
-    if np.isclose(hInnerEdge, 0):
-        hInnerEdge = 0.2 * max(outerRad) if np.any(np.isclose(innerRad, 0)) else 0.2 * max(innerRad)
-    if innerRad[0] > outerRad[0] or innerRad[1] > outerRad[1] or innerRad[2] > outerRad[2]:
-        ValueError("Inner ellipsoid does not fit inside outer ellipsoid")
-    # Create the two ellipsoid mesh using gmsh
-    gmsh.initialize()
-    gmsh.option.setNumber("General.Terminal", int(verbose))
-
-    gmsh.model.add("twoellipsoids")
-    # first add ellipsoid 1 of radius outerRad and center (0,0,0)
-    outer_ellipsoid = gmsh.model.occ.addSphere(0, 0, 0, 1.0)
-    gmsh.model.occ.dilate([(3, outer_ellipsoid)], 0, 0, 0, outerRad[0], outerRad[1], outerRad[2])
-    if np.any(np.isclose(innerRad, 0)):
-        # Use outer_ellipsoid only
-        gmsh.model.occ.synchronize()
-        gmsh.model.add_physical_group(3, [outer_ellipsoid], tag=outer_vol_tag)
-        facets = gmsh.model.getBoundary([(3, outer_ellipsoid)])
-        assert len(facets) == 1
-        gmsh.model.add_physical_group(2, [facets[0][1]], tag=outer_marker)
-    else:
-        # Add inner_ellipsoid (radius innerRad, center (0,0,0))
-        inner_ellipsoid = gmsh.model.occ.addSphere(0, 0, 0, 1.0)
-        gmsh.model.occ.dilate(
-            [(3, inner_ellipsoid)], 0, 0, 0, innerRad[0], innerRad[1], innerRad[2]
-        )
-        # Create interface between ellipsoids
-        two_ellipsoids, (outer_ellipsoid_map, inner_ellipsoid_map) = gmsh.model.occ.fragment(
-            [(3, outer_ellipsoid)], [(3, inner_ellipsoid)]
-        )
-        gmsh.model.occ.synchronize()
-
-        # Get the outer boundary
-        outer_shell = gmsh.model.getBoundary(two_ellipsoids, oriented=False)
-        assert len(outer_shell) == 1
-        # Get the inner boundary
-        inner_shell = gmsh.model.getBoundary(inner_ellipsoid_map, oriented=False)
-        assert len(inner_shell) == 1
-        # Add physical markers for facets
-        gmsh.model.add_physical_group(outer_shell[0][0], [outer_shell[0][1]], tag=outer_marker)
-        gmsh.model.add_physical_group(inner_shell[0][0], [inner_shell[0][1]], tag=interface_marker)
-
-        # Physical markers for
-        all_volumes = [tag[1] for tag in outer_ellipsoid_map]
-        inner_volume = [tag[1] for tag in inner_ellipsoid_map]
-        outer_volume = []
-        for vol in all_volumes:
-            if vol not in inner_volume:
-                outer_volume.append(vol)
-        gmsh.model.add_physical_group(3, outer_volume, tag=outer_vol_tag)
-        gmsh.model.add_physical_group(3, inner_volume, tag=inner_vol_tag)
 
     def meshSizeCallback(dim, tag, x, y, z, lc):
         # mesh length is hEdge at the PM (defaults to 0.1*outerRad,
@@ -323,28 +277,99 @@ def create_ellipsoids(
             lcTest = lc2 + (lc3 - lc2) * (1 - R_rel_inner)
         return lcTest
 
-    gmsh.model.mesh.setSizeCallback(meshSizeCallback)
-    # set off the other options for mesh size determination
-    gmsh.option.setNumber("Mesh.MeshSizeExtendFromBoundary", 0)
-    gmsh.option.setNumber("Mesh.MeshSizeFromPoints", 0)
-    gmsh.option.setNumber("Mesh.MeshSizeFromCurvature", 0)
-    # this changes the algorithm from Frontal-Delaunay to Delaunay,
-    # which may provide better results when there are larger gradients in mesh size
-    gmsh.option.setNumber("Mesh.Algorithm", 5)
+    import gmsh
 
-    gmsh.model.mesh.generate(3)
+    # Create temporary path for mesh (dependent on number of processes)
+    # Only generate mesh on rank 0
     rank = MPI.COMM_WORLD.rank
-    tmp_folder = pathlib.Path(f"tmp_ellipsoid_{outerRad}_{innerRad}_{rank}")
-    tmp_folder.mkdir(exist_ok=True)
+    size = MPI.COMM_WORLD.size
+    tmp_folder = pathlib.Path(f"tmp_ellipsoid_{outerRad}_{innerRad}_{size}")
+    if rank == 0:
+        tmp_folder.mkdir(exist_ok=True)
     gmsh_file = tmp_folder / "ellipsoids.msh"
-    gmsh.write(str(gmsh_file))
-    gmsh.finalize()
 
+    if rank == 0:
+        if np.any(np.isclose(outerRad, 0)):
+            ValueError("One of the outer radii is equal to zero")
+        if np.isclose(hEdge, 0):
+            hEdge = 0.1 * max(outerRad)
+        if np.isclose(hInnerEdge, 0):
+            hInnerEdge = (
+                0.2 * max(outerRad) if np.any(np.isclose(innerRad, 0)) else 0.2 * max(innerRad)
+            )
+        if innerRad[0] > outerRad[0] or innerRad[1] > outerRad[1] or innerRad[2] > outerRad[2]:
+            ValueError("Inner ellipsoid does not fit inside outer ellipsoid")
+        # Create the two ellipsoid mesh using gmsh
+        gmsh.initialize()
+        gmsh.option.setNumber("General.Terminal", int(verbose))
+
+        gmsh.model.add("twoellipsoids")
+        # first add ellipsoid 1 of radius outerRad and center (0,0,0)
+        outer_ellipsoid = gmsh.model.occ.addSphere(0, 0, 0, 1.0)
+        gmsh.model.occ.dilate(
+            [(3, outer_ellipsoid)], 0, 0, 0, outerRad[0], outerRad[1], outerRad[2]
+        )
+        if np.any(np.isclose(innerRad, 0)):
+            # Use outer_ellipsoid only
+            gmsh.model.occ.synchronize()
+            gmsh.model.add_physical_group(3, [outer_ellipsoid], tag=outer_vol_tag)
+            facets = gmsh.model.getBoundary([(3, outer_ellipsoid)])
+            assert len(facets) == 1
+            gmsh.model.add_physical_group(2, [facets[0][1]], tag=outer_marker)
+        else:
+            # Add inner_ellipsoid (radius innerRad, center (0,0,0))
+            inner_ellipsoid = gmsh.model.occ.addSphere(0, 0, 0, 1.0)
+            gmsh.model.occ.dilate(
+                [(3, inner_ellipsoid)], 0, 0, 0, innerRad[0], innerRad[1], innerRad[2]
+            )
+            # Create interface between ellipsoids
+            two_ellipsoids, (outer_ellipsoid_map, inner_ellipsoid_map) = gmsh.model.occ.fragment(
+                [(3, outer_ellipsoid)], [(3, inner_ellipsoid)]
+            )
+            gmsh.model.occ.synchronize()
+
+            # Get the outer boundary
+            outer_shell = gmsh.model.getBoundary(two_ellipsoids, oriented=False)
+            assert len(outer_shell) == 1
+            # Get the inner boundary
+            inner_shell = gmsh.model.getBoundary(inner_ellipsoid_map, oriented=False)
+            assert len(inner_shell) == 1
+            # Add physical markers for facets
+            gmsh.model.add_physical_group(outer_shell[0][0], [outer_shell[0][1]], tag=outer_marker)
+            gmsh.model.add_physical_group(
+                inner_shell[0][0], [inner_shell[0][1]], tag=interface_marker
+            )
+
+            # Physical markers for
+            all_volumes = [tag[1] for tag in outer_ellipsoid_map]
+            inner_volume = [tag[1] for tag in inner_ellipsoid_map]
+            outer_volume = []
+            for vol in all_volumes:
+                if vol not in inner_volume:
+                    outer_volume.append(vol)
+            gmsh.model.add_physical_group(3, outer_volume, tag=outer_vol_tag)
+            gmsh.model.add_physical_group(3, inner_volume, tag=inner_vol_tag)
+
+        gmsh.model.mesh.setSizeCallback(meshSizeCallback)
+        # set off the other options for mesh size determination
+        gmsh.option.setNumber("Mesh.MeshSizeExtendFromBoundary", 0)
+        gmsh.option.setNumber("Mesh.MeshSizeFromPoints", 0)
+        gmsh.option.setNumber("Mesh.MeshSizeFromCurvature", 0)
+        # this changes the algorithm from Frontal-Delaunay to Delaunay,
+        # which may provide better results when there are larger gradients in mesh size
+        gmsh.option.setNumber("Mesh.Algorithm", 5)
+
+        gmsh.model.mesh.generate(3)
+        gmsh.write(str(gmsh_file))
+        gmsh.finalize()
+    MPI.COMM_WORLD.Barrier()
     # return dolfin mesh of max dimension (parent mesh) and marker functions mf2 and mf3
     dmesh, mf2, mf3 = gmsh_to_dolfin(str(gmsh_file), tmp_folder, 3, comm)
     # remove tmp mesh and tmp folder
-    gmsh_file.unlink(missing_ok=False)
-    tmp_folder.rmdir()
+    if rank == 0:
+        gmsh_file.unlink(missing_ok=False)
+        tmp_folder.rmdir()
+
     # return dolfin mesh, mf2 (2d tags) and mf3 (3d tags)
     return (dmesh, mf2, mf3)
 
@@ -1289,37 +1314,40 @@ def gmsh_to_dolfin(
     """
     import meshio
 
-    # load, convert to xdmf, and save as temp files
-    mesh_in = meshio.read(gmsh_file_name)
-    if dimension == 2:
-        cell_type = "triangle"
-        facet_type = "line"
-    elif dimension == 3:
-        cell_type = "tetra"
-        facet_type = "triangle"
-    else:
-        ValueError(f"Mesh of dimension {dimension} not implemented")
-    # convert cell mesh
-    cells = mesh_in.get_cells_type(cell_type)
-    cell_data = mesh_in.get_cell_data("gmsh:physical", cell_type)  # extract values of tags
-    out_mesh_cell = meshio.Mesh(
-        points=mesh_in.points,
-        cells={cell_type: cells},
-        cell_data={"mf_data": [cell_data]},
-    )
     tmp_file_cell = tmp_folder / "tempmesh_cell.xdmf"
-    meshio.write(tmp_file_cell, out_mesh_cell)
-    # convert facet mesh
-    facets = mesh_in.get_cells_type(facet_type)
-    facet_data = mesh_in.get_cell_data("gmsh:physical", facet_type)  # extract values of tags
-    out_mesh_facet = meshio.Mesh(
-        points=mesh_in.points,
-        cells={facet_type: facets},
-        cell_data={"mf_data": [facet_data]},
-    )
     tmp_file_facet = tmp_folder / "tempmesh_facet.xdmf"
-    meshio.write(tmp_file_facet, out_mesh_facet)
 
+    if comm.rank == 0:
+        # load, convert to xdmf, and save as temp files
+        mesh_in = meshio.read(gmsh_file_name)
+        if dimension == 2:
+            cell_type = "triangle"
+            facet_type = "line"
+        elif dimension == 3:
+            cell_type = "tetra"
+            facet_type = "triangle"
+        else:
+            ValueError(f"Mesh of dimension {dimension} not implemented")
+        # convert cell mesh
+        cells = mesh_in.get_cells_type(cell_type)
+        cell_data = mesh_in.get_cell_data("gmsh:physical", cell_type)  # extract values of tags
+        out_mesh_cell = meshio.Mesh(
+            points=mesh_in.points,
+            cells={cell_type: cells},
+            cell_data={"mf_data": [cell_data]},
+        )
+        meshio.write(tmp_file_cell, out_mesh_cell)
+        # convert facet mesh
+        facets = mesh_in.get_cells_type(facet_type)
+        facet_data = mesh_in.get_cell_data("gmsh:physical", facet_type)  # extract values of tags
+        out_mesh_facet = meshio.Mesh(
+            points=mesh_in.points,
+            cells={facet_type: facets},
+            cell_data={"mf_data": [facet_data]},
+        )
+        meshio.write(tmp_file_facet, out_mesh_facet)
+
+    comm.Barrier()
     # convert xdmf mesh to dolfin-style mesh
     dmesh = d.Mesh(comm)
     mvc_cell = d.MeshValueCollection("size_t", dmesh, dimension)
@@ -1338,10 +1366,11 @@ def gmsh_to_dolfin(
     mf_facet.array()[np.where(mf_facet.array() > 1e9)[0]] = 0
 
     # remove temp meshes
-    tmp_file_cell.unlink(missing_ok=False)
-    tmp_file_cell.with_suffix(".h5").unlink(missing_ok=False)
-    tmp_file_facet.unlink(missing_ok=False)
-    tmp_file_facet.with_suffix(".h5").unlink(missing_ok=False)
+    if comm.rank == 0:
+        tmp_file_cell.unlink(missing_ok=False)
+        tmp_file_cell.with_suffix(".h5").unlink(missing_ok=False)
+        tmp_file_facet.unlink(missing_ok=False)
+        tmp_file_facet.with_suffix(".h5").unlink(missing_ok=False)
     # return dolfin mesh and mfs (marker functions)
     return (dmesh, mf_facet, mf_cell)
